Add arcade drive and differential drive getter methods to Drivetrain class
Added arcadeDrivePercentOutput method, which calls WPI lib arcade drive for driving the differential drivetrain in arcade mode with the joysticks.

Added getDifferentialDrive method which is a getter method for the raw differential drivetrain object. This might be useful in the future for path generation, although this functionality may be best handled within the class itself.