Connect joystick input to DriveStandard command, remove Drivetrain getDifferentialDrive() method
Xbox controllers are set up in RobotContainer as public and static, and the command DriveStandard now imports the RobotContainer class so it can access the joysticks directly. This cleans up the command calling process and the parameters that are used during calls.

Removed the getDifferentialDrive() method from Drivetrain subsystem class, as this object should inherently not be accessible directly. The goal of the subsystem class is to abstract away all hardware so that it is never directly addressed except by the class itself.