Update PID options for Drivetrain, comment out broken Shuffleboard code
Tested velocity-drive code for the first time. Did not work great.

The drivetrain during velocity-drive was shaking back and forth really quickly, which seemed to be caused by the encoders being inverted. Fixed by inverting them in Drivetrain.java constructor.

Shuffleboard code is configured incorrectly and will not compile, so commented out temporarily.

After inverting encoders, drivetrain runs but will not change velocity once activated. When velocity mode is turned on, the motors will go max speed in whatever direction is first applied by the joystick. The values may not be passed to the Talons correctly or at all, or may be received incorrectly?

See Also: #22, #16, #6, #14