Drivetrain: Re-implement NetworkTables, configure Shuffleboard, refactor constructor
These changes are not yet tested.

The previous implementation of NetworkTables for PIDF values caused the robot to fail to start, and therefore was temporarily removed in 6e1966b. This issue was caused because the NetworkTables PIDF values were being used as the default values for their own constructors before they were even initialized.

Additionally, the constructor for Drivetrain was getting long and gross, and more configuration keeps getting added.

Fixed the NetworkTables implementation, and fixed the way that Shuffleboard uses it.

Created private methods Drivetrain.config<category> which are called by Drivetrain(), to improve clarity and organization.

See Also:
- #6 (DS Display)
- #16 (Standardize Shuffleboard)
- [This Shuffleboard example](https://github.com/wpilibsuite/allwpilib/blob/main/wpilibjExamples/src/main/java/edu/wpi/first/wpilibj/examples/shuffleboard/Robot.java) for reference