Configure Shuffleboard subsystem indicators
For both competitions and tests, we want to have indicators for each subsystem on the Shuffleboard, as well as tools like Command selectors and PID controllers.

This commit adds indicators for each subsystem to the default Shuffleboard tab which resolves #6 (Dashboard Indicator), as well as a list of selectable commands which resolves #3 (Drive Mode Toggle).

Testing is still needed before #14 (Velocity Controlled Drive) can be closed.

See Also:
- #14 (Velocity Controlled Drive)