Create NetworkTables sync methods
NetworkTables values are utilized by each subsystem in order to integrate with the Shuffleboard dashboard. The values stored in NetworkTables need to be pulled into the hardware in order to enable control from the Shuffleboard, and values in the hardware need to be pushed to NetworkTables to enable indicators on the Shuffleboard.

Created a syncNetworkTables() method in each subsystem which is meant to run regularly during PeriodicTasks. This method pulls control data from Shuffleboard and pushes necessary hardware data back up. The idea is to keep the NetworkTables values up-to-date rather than rely on member variables. This applies to issue #23, and will also make #24, #18, and #20 easier to implement.