Move commands to subfolders, create Hopper collect/dispense commands
Moved commands to subfolders based on subsystem to improve organization. This will become more important as we create more complex actions.

Created new commands for collecting balls from the floor and for dispensing balls. Still needs to be configured to buttons, and needs specified heights/positions for dispensing and such.