Implement Hopper intake methods
Added an enum and Shuffleboard controls for the Hopper Intake. Intake can now be triggered with Hopper.setIntake(Intake), and use the enum to specify operating mode. As with other subsystem controls, the values are controlled over NetworkTables on the Shuffleboard, and default to those in Constants.java.

See Also:
- #31
- #20
- #23