Add getters/setters for Hopper Lift setpoints, add neutral override
The NetworkTables data for the Hopper subsystem is currently handled directly by the subsystem class, but there is currently no setters/getters for external commands or interfaces. Therefore the only way to change any of these values is directly through the Shuffleboard.

Additionally, there is no safety override of any kind for the Hopper subsystem.

This commit adds setters and getters for the Lift setpoints, which addresses #27 and #20.

This commit adds a method to the Hopper subsystem called setAllNeutral() which sets all motors to neutral output. This addresses #24, and should also be implemented on the Drivetrain subsystem.