Add Hopper Lift calibration command
When the robot boots up, the Hopper Lift does not always start in the same position nor does the encoder always return the right value. Additionally, if the encoder is jostled, the values can change.

Added a sequential command which runs the motor up and down to find the limits of its movement, and then resets the setpoint values for the Lift. This still needs testing.

See Also:
- #27