Refactor Hopper with safety override
The Hopper subsystem relies on a mag encoder for operating the Lift, which works great until it doesn't. We need a failsafe mode which operates the lift using basic percent output rather than the PID controller in case something goes wrong. This mode needs to be able to be toggled, and also should automatically toggle if the motor or encoder goes wonky.

Refactored the Hopper subsystem, and performed the following:
- Add default Lift setpoints from [2020 code](https://github.com/JHS-Viking-Robotics/FRC-2020)
- Store all Hopper state info in NetworkTables
- Create/rename setters/getters for NetworkTable state info. Methods no longer talk directly to motor controllers
- Add 'enable/disable Hopper subsystem' button to Shuffleboard Hopper tab
- Rearrange Shuffleboard Hopper tab
- Add 'subsystem enable' setter/getter
- Add subsystem motor safety check (checks safety config is enabled, current draw is not too high)
- Update NetworkTables sync method, also add subsystem safety check to Periodic Tasks

These changes have not been extensively tested, and all of the commands still need to be configured to buttons.

See Also:
- #30 (NetworkTables Reliability)
- #24 (Hopper manual override)