Add Hopper Manual override button to dashboard
The Hopper already has an automatic safety system which will disable the PID control on the Hopper, but now there is a manual override button on the Dashboard as well. This can be used in an emergency if the system never comes back online, or if it is not shutting itself off.

See Also: 24