Refactor command and command sequences organization
There are both one-off commands and command sequences for each subsystem. Moved the steps for the sequence into their own package so that they can be imported by SequenceCommand, and moved the SequenceCommands into the root folder so that all commands for a subsystem can be imported with one line.

See also: #35 (Hopper command structure)