Refactor Hopper setpoints and create a new TestMode
Hopper no longer uses NetworkTables values to update the setpoints for the lift and intake. Instead, there is a new TestMode command which simply lets the Lift get updated from the pre-existing setpoint NetworkTables slider.

Additionally, the enum for the Lift and Intake now has the values baked into it, so there are no more clunky getters for the setpoints.

See also: #35
Closes: #20