Refactor Hopper enums, Intake and Lift setters
The previous few commits reimplemented the manual and automatic overrides, but did not address how the motor controllers are actually updated.

Fixed this by adding a ControlMode parameter to the Intake and Lift enums. This way the public interface stays the same for using the Lift and Intake, but under the hood we can allow for disabling the controllers, and other modes like current control.

Also implemented current control of the Intake for holding balls in, although this will need testing so the motor doesn't overheat or catch fire.

See also: #31 (Intake current in), #35 (hopper commands), #24 (hopper override), #20 (hopper constants)