Configure Hopper test mode, add better comments
A testing command and a test mode were previously added to the Hopper subystem. This commit enables Hopper test mode whenever the Robot itself is put in test mode. This means there is one less button on the dashboard, and makes it less likely for test mode to be accidentally enabled.

See also: #33 (update docs), #35 (Hopper commands)