Add Robot-Characterization configuration, add testing gains to robot config
Added the configuration files for a RobotCharacterization project. The routine was ran, but did not seem to work the best due to the amount of drift on the robot. The trajectories that were tested did not work either. However, this is not likely due to the Characterization code itself being bad, rather it is the gains and the rest of the implementation.

See also: #40 (Characterize Robot)