Fix odometer, replace Trajectory following code with basic move code
The Pathweaver tutorial online is not made for Mecanum drivetrains, SparkMax's, CAN buses, or anything really. Could not get it to work, so instead made a basic movement command for auton. Much simpler and seems to work well.

Fixed a bug where the odometer/speedometer were using the wrong units. At one point the odometer was accurate, so I think there must've been an update to the SparkMax API recently.