Fix missing Main and vendordeps
A few times now the Main.java, Robot.java, and a few of the vendordeps files have disappeared from git. I'm not really sure why this is. I think it has to do with when we reset the repo for the 2022 season and started making a bunch of branches, it got deleted originally and then that permeated throughout each branch. Build is successful now at least.